feat(agent): Scan environment for label configuration (#467)
Scan environment for singular and plural label configuration

Added new global variable env_labels to keep track of the labels that were extracted from the environment. In particular, labels are designated with the key NEW_RELIC_LABELS or by keys prefixed with NEW_RELIC_LABEL_
Updated the environment variable parsing logic to handle both looking for a specific key and to handle parsing keys with prefixes into strings.
Integrated environment labels with the ini labels to take advantages of all the labels functionality. By appending the environment variables to the end of the ini string, we ensure the environmental variables take precedence when it is entered into the label hash in nr_php_txn_get_labels which is then associated with the transaction. Additionally, with the environmental and ini values combined, we are able to take advantage of all the previous label structures, tests, and string validation (nr_labels_parse).