feat(daemon): Implement log event limits configuration
Implements changes to support accepting a log event limit
from the agent (from INI config) and passing it in a
collector connect request.  The collector response is
parsed and the collector suggested value for the log limit
it retrieved and compared to the agent requested log limit.
The smaller of these values is then used for setting up
new harvest cycles.