fix(agent): Enforce smallest log event limit in agent
Need to enforce the smallest of the collector harvest limit and the agent
config INI value here.  This is in part because the enforcement in the daemon
will not be exercised in integration tests due to how the integration_runner
doesn't follow the same code path when processing a connect response
as in the normal operation of the daemon.