test(logging): Fix monolog tests for priority sampling ordering
Now log events are sampled using priority sampling the order
of the events in the harvest will not necessarily be chronological
but will follow the order in the minmax heap.  Due to the limited
resolution of the timestamp stored on log messages (milliseconds)
it was necessary to inject a delay between log messages to
make sure the age component of the comparison function used
to insert events into the heap reflected the actual ages.  Otherwise
the order output can be variable as to whether all the log
messages happened in the same millisecond or not.