fix(logging): Improve efficiency of log event -> JSON
A new method nr_log_event_to_json_buffer_ex() is created which allows
disabling the wrapping "[" "]" on the JSON string.  These chararacters
are required for the resulting JSON to be valid standalone.  However,
to improve the efficiency of sending the log harvest payload it is
advantageous to leave these characters off as the log events JSON
strings are joined in a large JSON list in the collector payload.  If these
characters were present it would require extracting the substring w/o
them for each log event harvested.