fix(agent): Improve nr_log_events_add_event() efficiency
Changes the API for nr_log_events_add_event() to no longer clone the event
passed in - now it takes ownership of the log event and is responsibe for
freeing the memory.   Should help reduce the amount of memory copied
around especially if there is a large amount of log message data.  This
change does mean that now nr_log_events_add_event() will only work if
the event passed in was allocated with nr_log_event_create().

This enhancement required changing only one place in the agent code
in nr_txn.c but quite a few places in the unit tests.  The new code passes these
with valgrind.