feat(agent, daemon): Add priority sampled log event harvest  (#515)
* fix(test): Allow larger EXPECT strings for tests

* feat(agent): Add API to retrive segment priority flag

* test(agent): Update nr_segment_t unit test

* feat(agent): Extend nr_log_event_t for priority sampling

* Log records are now limited to 32768 characters to follow spec
* Added nr_log_event_clone()
* Added priority field to nr_log_event_t to allow priority sampling
  of log events
* Added nr_log_event_set_priority() API call
* Removed extra nr_free() in nr_log_event_destroy()
* Changed JSON creation to use "level" instead of "log.level"
* Brought nr_log_event_t unit tests up to date

* feat(agent): Implement priority sampling for log events

* Log events now stored on minmax heap like span events
* Log events are stored as type nr_log_event_t , not as JSON
  strings as before
* Segments with linked log events have priority flag
  NR_SEGMENT_PRIORITY_LOG set (although this  was
  probably happening due to call to nr_txn_get_current_span_in() in
  log_event_set_linking_metadata()
* The API call nr_log_events_get_event_json() was  removed as it didn't
  map to the minmax heap representation of log events.  Now it is
  necessary to call nr_log_events_to_vector() with a pre-allocated
  nr_vector_t to access log events.
* Unit tests for log events updated and extended to match changes
* Agent flatbuffer creation code extends to send log events

* fix(daemon): Add custom JSON creation for log events

* fix(daemon): Add log event harvest period to harvest

Added log harvest period to consideration of  if all periods are the same.

* fix(test): Update processor test for log event JSON

* test(logging): Added logging integration tests

* test(logging): Add tests for high security mode enabled

* fix(logging): Improve efficiency of log event -> JSON

A new method nr_log_event_to_json_buffer_ex() is created which allows
disabling the wrapping "[" "]" on the JSON string.  These chararacters
are required for the resulting JSON to be valid standalone.  However,
to improve the efficiency of sending the log harvest payload it is
advantageous to leave these characters off as the log events JSON
strings are joined in a large JSON list in the collector payload.  If these
characters were present it would require extracting the substring w/o
them for each log event harvested.

* test(daemon): Update harvest trigger test to include log event

* fix(daemon): Fixes logging supportability metric names

* fix(agent): Creates log event JSON for valid fields only

* chore(agent): Fix typos in cmd_txndata_transmit.c

Co-authored-by: Amber Sistla <ambersistla@gmail.com>

* style(agent): Consistently use braces on if statements

* chore(style): Remove unneeded comment/include

Co-authored-by: Amber Sistla <ambersistla@gmail.com>

* chore(agent): Fix typos

Co-authored-by: Amber Sistla <ambersistla@gmail.com>

* chore(agent): Fix typos

Co-authored-by: Amber Sistla <ambersistla@gmail.com>

* chore(agent): Fix typos

Co-authored-by: Amber Sistla <ambersistla@gmail.com>

* chore(agent): Fix typos

Co-authored-by: Amber Sistla <ambersistla@gmail.com>

* chore(agent): Fix comment style to match guide

Co-authored-by: Amber Sistla <ambersistla@gmail.com>

* fix(agent): Make check for NULL case more clear

* style(agent): Consistently use braces on if statements

* chore(agent): Removed unneeded include

Co-authored-by: Amber Sistla <ambersistla@gmail.com>

* chore(agent): Make comment formatting consistent with guide

Co-authored-by: Amber Sistla <ambersistla@gmail.com>

* chore(agent): Make function docs more clear

* fix(agent): Initialize local variable

Co-authored-by: Amber Sistla <ambersistla@gmail.com>

* fix(agent): Fix memory leak

* fix(agent): Improve nr_log_events_add_event() efficiency

Changes the API for nr_log_events_add_event() to no longer clone the event
passed in - now it takes ownership of the log event and is responsibe for
freeing the memory.   Should help reduce the amount of memory copied
around especially if there is a large amount of log message data.  This
change does mean that now nr_log_events_add_event() will only work if
the event passed in was allocated with nr_log_event_create().

This enhancement required changing only one place in the agent code
in nr_txn.c but quite a few places in the unit tests.  The new code passes these
with valgrind.

* fix(agent): Remove unneeded function and tests

* chore(docs): Improve API documentation

* fix(agent): Rework code to reduce private includes required

* fix(agent): Improve robustness of nr_log_events_is_sampling()

* chore(agent): Fix typos in function documention

Co-authored-by: Michal Nowacki <mnowacki@newrelic.com>

* chore(agent): clang format prevention

Co-authored-by: Michal Nowacki <mnowacki@newrelic.com>

* chore(agent): Prevent clang formatting

Co-authored-by: Michal Nowacki <mnowacki@newrelic.com>

* fix(agent): Improve variable name meaning

* fix(agent): Improve variable name meaning

* test(agent): Improve robustness of test

The minmax heap implementation does not guarantee
the order the iterator function will traverse the heap.
This change sorts the resulting vector so we can reliably
compare the two events remaining in the heap to what
we expect.

* fix(agent): Remove unneeded log event functionality

This commit removes the function nr_log_event_to_json_buffer_ex()
as it is not needed.  Originally the idea was to have a function to
produce a more compact JSON representation but after review
the compact representation is adequate for all use cases.  This
makes the function nr_log_event_to_json_buffer_ex() unnecessary.

Code using this function as well as affected test cases have been
updated.

* test(agent): Improve log events test

Due to the uncertainty of the order events pulled from a
minmax heap via an iterator some test cases needed to
be reworked.  Instead of testing the internal construction
of the minmax heap the test case verifies the correct
events have been inserted.
The operation of the minmax heap is currently better
exercised in the test_events_sample() test which inserts
more events in the heap than it can hold and verifies
the highest priority events are retained.

Co-authored-by: Amber Sistla <ambersistla@gmail.com>
Co-authored-by: Michal Nowacki <mnowacki@newrelic.com>